refactor repo structure, adjust Dockerfiles apropriatly
