+++ Dockerfile.signature-collection
@@ -9,7 +9,7 @@ COPY pyproject.toml poetry.lock* ./
 RUN poetry config virtualenvs.create false \
   && poetry install --no-interaction --no-ansi
 
-COPY ./collection ./collection
+COPY ./src ./src
 COPY ./db ./db
 
-CMD ["python3", "collection/historical_signatures/signature_collector.py"]
+CMD ["python3", "src/collection/historical_signatures/signature_collector.py"]
