+++ src/collection/shared/generic_collector.py
